+++ src/Phan/Language/Element/Func.php
@@ -333,11 +333,12 @@ class Func extends AddressableElement implements FunctionInterface
     /** @return array{0:string,1:string} [string $namespace, string $text] */
     public function toStubInfo() : array
     {
+        $fqsen = $this->getFQSEN();
         $stub = 'function ';
         if ($this->returnsRef()) {
             $stub .= '&';
         }
-        $stub .= $this->getName();
+        $stub .= $fqsen->getName();
         $stub .= '(' . implode(', ', array_map(function (Parameter $parameter) : string {
             return $parameter->toStubString();
         }, $this->getRealParameterList())) . ')';
@@ -347,7 +348,7 @@ class Func extends AddressableElement implements FunctionInterface
 
         $stub .= ' {}' . "\n";
 
-        $namespace = ltrim($this->getFQSEN()->getNamespace(), '\\');
+        $namespace = ltrim($fqsen->getNamespace(), '\\');
         return [$namespace, $stub];
     }
 }